Corrected missing flags for TRR packet sending and receiving

+++ model/aodv-packet.h
@@ -342,7 +342,7 @@ class TRRHeader : public Header
 {
 public:
   /// c-tor
-	TRRHeader (double GT = 0, double DT = 0,
+	TRRHeader (uint32_t GT = 0, uint32_t DT = 0,
               uint32_t trrID = 0, Ipv4Address dst = Ipv4Address (),
               uint32_t dstSeqNo = 0, Ipv4Address origin = Ipv4Address (),
               uint32_t originSeqNo = 0);
@@ -358,10 +358,10 @@ public:
 
   ///\name Fields
   //\{
-  double GetDT () {return m_DT;}
-  void setDT(double dt) { m_DT = dt; }
-  double GetGT () {return m_GT;}
-  void setGT(double gt) { m_GT = gt; }
+  uint32_t GetDT () {return m_DT;}
+  void setDT(uint32_t dt) { m_DT = dt; }
+  uint32_t GetGT () {return m_GT;}
+  void setGT(uint32_t gt) { m_GT = gt; }
   void SetId (uint32_t id) { m_trrID = id; }
   uint32_t GetId () const { return m_trrID; }
   void SetDst (Ipv4Address a) { m_dst = a; }
@@ -390,8 +390,8 @@ public:
 
   bool operator== (TRRHeader const & o) const;
 private:
-  double 		 m_GT;			   ///<Global Trust
-  double 		 m_DT;			   ///<Direct Trust
+  uint32_t 		 m_GT;			   ///<Global Trust
+  uint32_t 		 m_DT;			   ///<Direct Trust
   uint32_t       m_trrID;          ///< TRR ID
   Ipv4Address    m_dst;            ///< Destination IP Address
   uint32_t       m_dstSeqNo;       ///< Destination Sequence Number
